Rename site-type-route-supports to site-type-feature-support
